Added ImGui component
The ImGui component shows 2 exercises in which they calculate time to edit objects with different object size. The first exercise shows the time graph of a vector of ints thats gets edited all then edited halve as many and another halve as many until 1/1024.
The second exercise does the same for 2 different object sizes one with 68 byte size and another with 16 bytes size. This exercise also shows the difference between the two.

+++ Minigin/GameObject.h
@@ -63,7 +63,7 @@ namespace dae
 		void SetPositionDirty()
 		{
 			m_Transform.SetPositionDirty();
-			for (int i{}; i < m_pChildren.size(); ++i)
+			for (uint16_t i{}; i < m_pChildren.size(); ++i)
 			{
 				m_pChildren[i]->SetPositionDirty();
 			}
